+++ CHANGELOG.md
@@ -7,3 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [1.0.15] 2022.11.24
 ### Fixed
 - Nerf drops when killed by none players
+- Don't add spawns if they are disabled
+- Some fixes to Spawn rules
+- Update Arch dependency
+### Added 
+- CreeperTypes json now allows biomesTags list for handling spawns